+++ Elements/src/Geometry/Polygon.cs
@@ -55,6 +55,16 @@ namespace Elements.Geometry
         }
 
         /// <summary>
+        /// Construct a transform in the plane of this polygon, with its Z axis along the polygon's normal.
+        /// </summary>
+        /// <returns></returns>
+        public Transform ToTransform()
+        {
+            var normal = Normal();
+            return new Transform(Vertices[0], Vertices[1] - Vertices[0], normal);
+        }
+
+        /// <summary>
         /// Tests if the supplied Vector3 is within this Polygon without coincidence with an edge when compared on a shared plane.
         /// </summary>
         /// <param name="vector">The Vector3 to compare to this Polygon.</param>
@@ -362,14 +372,10 @@ namespace Elements.Geometry
         /// <param name="pl">The polyline with which to split.</param>
         public List<Polygon> Split(Polyline pl)
         {
-            var plXform = this.Vertices.ToTransform();
+            var plXform = this.ToTransform();
             var inverse = new Transform(plXform);
             inverse.Invert();
             var thisInXY = this.TransformedPolygon(inverse);
-            if (thisInXY.IsClockWise())
-            {
-                thisInXY = thisInXY.Reversed();
-            }
             // Construct a half-edge graph from the polygon and the polyline
             var graph = Elements.Spatial.HalfEdgeGraph2d.Construct(thisInXY, pl.TransformedPolyline(inverse));
             // Find closed regions in that graph
@@ -382,14 +388,10 @@ namespace Elements.Geometry
         /// <param name="polylines">The polylines with which to split.</param>
         public List<Polygon> Split(IEnumerable<Polyline> polylines, Model m = null)
         {
-            var plXform = this.Vertices.ToTransform();
+            var plXform = this.ToTransform();
             var inverse = new Transform(plXform);
             inverse.Invert();
             var thisInXY = this.TransformedPolygon(inverse);
-            if (thisInXY.IsClockWise())
-            {
-                thisInXY = thisInXY.Reversed();
-            }
             // Construct a half-edge graph from the polygon and the polylines
             m?.AddElement(this.TransformedPolygon(inverse));
             m?.AddElements(polylines.Select(p => new ModelCurve(p.TransformedPolyline(inverse))));