+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Added
 
+- `Polygon.ToTransform()`
 - `Grid2d.IsOutside()`
 
 ### Removed
+
 - `BuiltInMaterials.Dirt`
 - `BuiltInMaterials.Grass`
 
@@ -16,12 +18,12 @@
 - `ConstructedSolid` serializes and deserializes correctly.
 - `Solid.AddFace(Polygon, Polygon[])` can take an optional third `mergeVerticesAndEdges` argument which will automatically reuse existing edges + vertices in the solid.
 - Adds optional `tolerance` parameter to `Line.ExtendTo(Polygon)`, `Line.ExtendTo(IEnumerable<Line>)`, `Vector3.IsParallelTo(Vector3)`.
+- `Polygon.Split()` can now handle polygons that are not in the XY plane.
 
 ### Fixed
 
 - Fixed a bug in `ConvexHull.FromPoints` when multiple X coordinates are equal.
 - Fixed a bug in `Grid2d(Polygon, Vector3, Vector3, Vector3)` where U or V directions skew slightly when they nearly parallel with a boundary edge.
-
 
 ## 0.8.5